refactor(thoth): fold vite-plugin in as thoth/vite
Moves vite-plugin → packages/thoth/vite/ (plugin imports thoth via ../compiler.js); thoth exports ./vite and absorbs @rollup/pluginutils + a vite peer-dep. azoth's ./vite-plugin export now re-exports @azothjs/thoth/vite — export PATH unchanged, so NOT breaking. Root vite.config.js imports the plugin from thoth/vite; the node-test exclude is now a cwd-robust glob (**/vite/index.test.js) so per-package runs don't trip on it. vite-plugin package removed; lock regenerated (pinned pnpm).

maya 137, valhalla 70, thoth 12960; azoth/vite-plugin resolves to the plugin fn. (vite-plugin test stays excluded — its brittleness is the parked vite-test rethink.)
