+++ packages/thoth/package.json
@@ -25,12 +25,21 @@
             "import": {
                 "default": "./compiler.js"
             }
+        },
+        "./vite": {
+            "import": {
+                "default": "./vite/index.js"
+            }
         }
     },
     "scripts": {
         "test": "vitest"
     },
+    "peerDependencies": {
+        "vite": "^5.0.12"
+    },
     "dependencies": {
+        "@rollup/pluginutils": "^5.1.0",
         "acorn": "^8.11.3",
         "acorn-jsx": "^5.3.2",
         "aria-attributes": "^2.0.1",
