+++ packages/thoth/vite/TSX-SUPPORT.md
