--- packages/vite-plugin/CHANGELOG.md
@@ -1,9 +1,0 @@
-# @azothjs/vite-plugin
-
-## 0.4.5
-
-### Patch Changes
-
-- Fix multi-page build template injection. Module configuration and dependency updates.
-- Updated dependencies
-  - @azothjs/thoth@0.4.5
