--- packages/vite-plugin/package.json
@@ -1,42 +1,0 @@
-{
-    "name": "@azothjs/vite-plugin",
-    "version": "0.4.5",
-    "description": "azoth plugins for vite",
-    "author": "Marty Nelson",
-    "license": "MIT",
-    "keywords": [
-        "vite-plugin",
-        "vite plugin",
-        "vite",
-        "azoth",
-        "azoth plugins"
-    ],
-    "engines": {
-        "node": "^18.0.0 || >=20"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/azothjs/azoth.git",
-        "directory": "packages/vite-plugin"
-    },
-    "type": "module",
-    "files": [
-        "index.js",
-        "dist"
-    ],
-    "exports": {
-        ".": {
-            "import": {
-                "default": "./index.js"
-            }
-        }
-    },
-    "peerDependencies": {
-        "vite": "^5.0.12"
-    },
-    "dependencies": {
-        "@azothjs/thoth": "workspace:*",
-        "@rollup/pluginutils": "^5.1.0",
-        "source-map": "^0.7.4"
-    }
-}
